Rename top-level Django project directories
    * Project modules will now end with `_project`
